Refactor configuration dialog to use dedicated settings widgets for encoding, processing, and advanced options; enhance UI organization and maintainability.
